chore(admin): rename component for clarity
